requirements.txt
tinyfish==0.2.4
fastapi==0.115.14
uvicorn[standard]==0.34.3
openai==1.54.4
httpx==0.27.2
beautifulsoup4==4.12.3
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.6.1
aiofiles==24.1.0
python-multipart==0.0.12
sqlalchemy==2.0.36
aiosqlite==0.20.0
alembic==1.14.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
openpyxl==3.1.5
# For PostgreSQL: swap aiosqlite for asyncpg and update DATABASE_URL in .env
# asyncpg==0.30.0
